move VQH class to core/
